forgot

+++ src/Tiller.h
@@ -11,17 +11,17 @@ class Tiller {
 
 public:
 	Tiller() {};
-	Tiller(BBox bbox_, KDnode* kd_, float radius, string rlt): bbox(bbox_),kd_tree(kd_), points_in_grid(1.0 / (2 * radius) ), result_dir(rlt){ }
+	Tiller(BBox bbox_, KDnode* kd_, Float radius, string rlt): bbox(bbox_),kd_tree(kd_), points_in_grid(1.0 / (2 * radius) ), result_dir(rlt){ }
 	
 	~Tiller() {};
-	void insert_in_gap(Point2D pivotPoint, Point2D conflictPoint, double radius, int depth);
-	void eliminate_for_maximal(double radius);
-	void process_pivot_point(Point2D& pivotPoint, int pri, double radius, int grid_idx, int ingrid_idx, int chain);
-	void DivideConquerTiling(BBox bbox, double r, int axis, float ratio);
-	void tilePoints(BBox bbox,Point2D offset, float ratio){
+	void insert_in_gap(Point2D pivotPoint, Point2D conflictPoint, Float radius, int depth);
+	void eliminate_for_maximal(Float radius);
+	void process_pivot_point(Point2D& pivotPoint, int pri, Float radius, int grid_idx, int ingrid_idx, int chain);
+	void DivideConquerTiling(BBox bbox, Float r, int axis, Float ratio);
+	void tilePoints(BBox bbox,Point2D offset, Float ratio){
 		traverse_and_classify(kd_tree,bbox, offset, ratio);
 	}
-	void traverse_and_classify(KDnode* root, BBox query, Point2D offset, float ratio);
+	void traverse_and_classify(KDnode* root, BBox query, Point2D offset, Float ratio);
 
 	int getGridCount() { return points_in_grid.getCount(); }
 
@@ -29,8 +29,8 @@ public:
 	{
 		points_in_grid.printToFile(output_file_name);
 	}
-	void test_maximal(double radius);
-	void test_conflict(double radius);
+	void test_maximal(Float radius);
+	void test_conflict(Float radius);
 	BBox bbox;
 	KDnode* kd_tree;
 	UniformGrids points_in_grid;
